Move search controller
