+++ server_files/controllers/getReply.js
@@ -1,0 +1,12 @@
+function getReply(results) {
+  try {
+    return results.length > 0 ? {data: results} : {message: 'No results found.'};
+  } catch (err) {
+    console.log(err);
+    return {message: 'Server error!'};
+  }
+}
+
+export {
+  getReply
+};
