+++ server_files/routes/routes.js
@@ -1,6 +1,9 @@
 import express from 'express';
 import sequelize from 'sequelize';
 import db from '../database/initDB.js';
+import { getAllIncidents } from '../controllers/incidents.js';
+import { getSearchResults } from '../controllers/search.js';
+
 const Op = sequelize.Op;
 const router = express.Router();
 
@@ -33,14 +36,7 @@ router.route('/newRoute')
 // INCIDENTS
 router.route('/incidents')
   .get(async (req, res) => {
-    try {
-      const all_incidents = await db.incidents.findAll();
-      const reply = getReply(all_incidents);
-      res.json(reply);
-    } catch (err) {
-      console.error(err);
-      res.send('Server Error!');
-    }
+    await getAllIncidents(req, res);
   })
   .post(async (req, res) => {
     try {
@@ -505,73 +501,7 @@ router.route('/units/:unit_id')
 
 router.route('/search')
 .get(async (req, res) => {
-  try {
-    // console.log(req)
-    let replyData = []
-    const matchCalls = await db.calls.findAll({
-      where: {
-        [Op.or]: [
-          {
-            call_type: {
-              [Op.startsWith]: req.query.queryText
-            } 
-          },
-          {
-            call_class: {
-              [Op.startsWith]: req.query.queryText
-            }
-          }
-        ]
-      }
-    });
-    let callsToIncidents;
-    if (matchCalls !== 'undefined') {
-      const call_ids = matchCalls.map((call) => {
-        // console.log(call.dataValues.call_id);
-        return call.dataValues.call_id;
-      });
-      callsToIncidents = await db.incidents.findAll({
-        where: {
-          call_id: {
-            [Op.in]: call_ids
-          }
-        }
-      })
-    }
-    
-    console.log(`${req.query.endDate} 23:59:59`)
-    const startDate = Date.parse(`${req.query.startDate}T00:00:00-00:00`);
-    const endDate = Date.parse(`${req.query.endDate}T23:59:59-00:00`);
-
-    const matchIncidents = await db.incidents.findAll({
-      where: {
-        [Op.and]: [{
-          [Op.or]: [
-            { description: {
-              [Op.startsWith]: req.query.queryText
-            } },
-            { postal_code: {
-              [Op.startsWith]: req.query.queryText
-            } },
-          ]},
-          {
-            date: {
-              [Op.between]: [startDate, endDate]
-            }
-          }
-      ]}
-    });
-    console.log(matchIncidents);
-    const incidentData = matchIncidents + callsToIncidents;
-    // console.log(incidentData);
-    replyData += [{'incidents': incidentData}];
-    // console.log(replyData);
-    const reply = getReply(replyData);
-    res.json(reply);
-  } catch (err) {
-    console.error(err);
-    res.send('Server Error!');
-  }
+  await getSearchResults(res, req);
 })
 .post(async (req, res) => {
   res.send('Action not available.');
